OpenStarkTFW/Datahub/DatahubAPI/requirements.txt
tornado
munch
tormysql
pymysql
pycryptodome
tornado_smtp
exchangelib
phantomas
win_unicode_console
paramiko
openpyxl
uvloop
python-jenkins
https://github.com/TronGeek/robotframework-QTLibrary/archive/master.zip